Fix monitoring thread early exit in production.
Incredibly simple bug, but it took forever to isolate this. The monitoring thread in a ProcessRunner actor checks on an external OS process. This thread was checking if our internal status for the process was 'Running', but that status is set immediately after the monitoring thread is started.  When running on multiple cores in production, the monitoring thread would start running either with stale data, or before the status was updated.

The quick fix was just to run the monitoring thread if the status is less than or equal to running. However, to be safe, I've also added a guard which checks if we're stalled in a 'NotStarted' state. That shouldn't happen, but it removes the possibility of leaking threads if for some reason it does.